Move GTFS meta helpers to common module
